- move LockScreenViewModel.kt to common module - handle reset and close on autologout for FancyAlertDialog.kt
